+++ runtime/src/int.rs
@@ -17,7 +17,7 @@ impl BlockPtr {
 
     pub fn new_int(x: i64) -> BlockPtr {
         let x: BlockPtr = unsafe { mem::transmute(x) };
-        Self::new(KnownTag::Int, [x])
+        Self::new(KnownTag::Int as u16, [x])
     }
 }
